Add StrInt
A simple class (and a base class which will be required by all Str
Construct objects) to handle build and parse. Currently, the parsing
uses the Parse package (https://pypi.org/project/parse/). But it
appears to have some problems when compared to the formats supported
by the `format` method. The corresponding test cases are marked
with xfail.

Also, added the minimal set of files to run unit tests by pytest.