Add StrStruct
Adding the basics of StrStruct. StrStruct is basically a list of other
StrConstruct objects (including StrStruct itself) that can be built
when the parent StrStruct is built.

Overriding the division operators is used as a means to name the
StrConstructs, while providing a user-friendly syntax for user (similar
to Construct).