Add StrDefault (#6)
StrDefault, similar to Default in Construct, receives a construct
object and a default value. When building without a value, the default
value will be passed on the provided construct. Building with a value
is exactly similar to building the underlying construct with that
value.

The default value has no impact on parsing.