A more strict parse algorithm for StrInt
There are two restrictions that were missing from the parse
library; case sensitivity for hex values (in fact "X" is not
a valid format for parse), and length checking.

With the new parse method, if length is provided in the format,
only that many characters from the input will be consumed. If
no length is provided, the algorithm consumes as many characters
compatible with the format as possible.