StrStruct can parse even without a separator (#11)
All the strConstruct objects now perform a greedy search if their
format does not specify length. If it does, they consume only as
many characters as specified in length. The greedy search continues
as long the found characters form a parsable string. As soon as
a non-parsable character is found (e.g. "," by a StrInt object),
the parse algorithmm stops.

As such, they needed a new API to return the remaining string
that their parse method did not consume. With that approach,
StrStruct object can also work without separator characters.