Add StrSwitch
StrSwitch is able to work with normal values as well as callables
to determing the required build/parse path.

Also fixed a corner case in StrStruct. The named values need to be
able to build, if they can, when no key-value pair is provided.
Example str-construct classes for this scenario are StrDefault and
StrConst.