Maintaine context while building and parsing
In addition to getting arguments in the build and parse methods,
each field, after being parsed or built, is now added to the context.
This makes it possible to refer to the fields inside a StrStruct in
other fields. One common example is StrSwitch, where we want to
"switch" based on the value of another field.

The context is a dictionary. Every time a value is successfully
built in StrStruct, it is added to context. Note that in this case
the value is what is given to the build method (not its output).

Similarly, every time a field is parsed, it is added to the context.
But there is a difference. This time the output of the parse function
is added to the context and not its argument. This makes it possible
to refer to other fields always by their names, regardless of
whether it's building or parsing.